Use serializer to determine relationship IDs when possible
The `Relationship` class currently has two possible methods for determining the ID for a relationship:
1. It can use the `serializer` associated with the relationship.
2. It has an option called `id_method_name`, where `id_method_name` is the name of a method on the parent record that can used to look up the ID of the related record without loading the related record.

The `id_method_name` is not a reliable method of determining IDs for a couple of reasons:
1. A value is not always provided and the calculated default is not always accurate.
2. The parent doesn't necessarily always have a method for finding the related ID directly, and there is no guarantee that it will be performant.
3. If the method does not exist, the user will get a not very helpful error message.

Since a serializer is always the source of truth for determining how a record gets serialized, it would generally make sense to use that where possible.

This PR updates the `Relationship` class to use the serializer class to calculate IDs in cases where an `id_method_name` is not provided and the related record and serializer are both available.